+++ xtra-libp2p-offer/src/current.rs
@@ -1,0 +1,5 @@
+pub mod maker;
+mod protocol;
+pub mod taker;
+
+pub const PROTOCOL: &str = "/itchysats/offer/2.0.0";