+++ xtra-libp2p-offer/src/deprecated.rs
@@ -1,0 +1,4 @@
+pub mod maker;
+pub mod protocol;
+
+pub const PROTOCOL: &str = "/itchysats/offer/1.0.0";